requirements.txt
absl-py==0.7.1
appnope==0.1.0
array2gif==1.0.4
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.6
bleach==3.1.0
blis==0.2.4
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colour==0.1.5
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
dtreeviz==0.4
entrypoints==0.3
filelock==3.0.12
findspark==1.3.0
future==0.17.1
gast==0.2.2
graphviz==0.11
grpcio==1.20.0
h5py==2.9.0
hyperas==0.4.1
hyperopt==0.1.2
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
llvmlite==0.29.0
logger==1.4
lucid==0.3.8
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
more-itertools==7.1.0
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numba==0.44.1
numpy==1.16.3
opencv-contrib-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.4.0
patsy==0.5.1
pbr==5.1.3
PDPbox==0.2.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.2
ptyprocess==0.6.0
public==2019.4.13
py4j==0.10.7
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
pymongo==3.8.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyparsing==2.4.0
pyrsistent==0.15.1
pyspark==2.4.3
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
self==2019.4.13
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
spacy==2.1.6
srsly==0.0.7
statsmodels==0.9.0
storage==0.0.4.3
tensorboard==1.13.1
tensorboardX==1.7
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
thinc==7.0.8
torch==1.1.0.post2
torchvision==0.3.0
tornado==6.0.2
tqdm==4.32.2
traitlets==4.3.2
umap-learn==0.3.9
urllib3==1.25.3
utils==0.9.0
virtualenv==16.7.0
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
xlrd==1.2.0
